Changed datatype of cnt from int to long long

+++ tests/definitions/perft.cpp
@@ -19,7 +19,7 @@ struct PerftTest : public ::testing::Test {
     }
 };
 
-int cnt = 0;
+ll cnt = 0;
 
 void setup() {
     initBitMasks();
